#23 move components to app folder
